# openapi.yaml
openapi: 3.1.0
info:
  title: DevMate Core API Specification
  description: |
    Production-ready REST and RPC API contract for the DevMate Telegram-based Personal Operating System platform.
    Defines schemas, endpoints, error response wrappers, webhooks, and callbacks for delivery adapters and future clients.
  version: 1.0.0
  termsOfService: https://devmate.internal/terms
  contact:
    name: DevMate API Platform Team
    email: platform-api@devmate.internal
    url: https://devmate.internal/support
  license:
    name: Proprietary License
    url: https://devmate.internal/license
...
servers:
  - url: https://api.devmate.internal/v1
    description: Production Server Gateway
  - url: https://staging-api.devmate.internal/v1
    description: Staging Environment Server
  - url: https://dev-api.devmate.internal/v1
    description: Development Integration Server
  - url: http://localhost:8080/v1
    description: Local Developer Container Server
security:
  - BearerAuth: []
paths:
  /auth/telegram:
    post:
      summary: Authenticate with Telegram
      description: Validates Telegram login credentials hash signature and returns a session access token.
      operationId: authenticateTelegram
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramAuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Issues a new short-lived access token using the refresh token stored in secure cookies.
      operationId: refreshAccessToken
      tags:
        - Authentication
      security:
        - RefreshTokenCookie: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/logout:
    post:
      summary: Terminate session
      description: Invalidates active access and refresh tokens, terminating the user session.
      operationId: logoutSession
      tags:
        - Authentication
      responses:
        '204':
          description: Session terminated successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /reminders/{id}:
    get:
      summary: Get reminder details
      description: Retrieves a single reminder schedule configuration.
      operationId: getReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reminder details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update reminder trigger details
      description: Modifies text, recurrence rules, or target trigger timestamps.
      operationId: updateReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Reminder rescheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Remove reminder alert
      description: Soft deletes a reminder schedule.
      operationId: deleteReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Reminder removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /calendar/events:
    get:
      summary: Get calendar events
      description: Retrieve list of scheduled events between two date thresholds.
      operationId: getCalendarEvents
      tags:
        - Calendar
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Calendar events list compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create calendar event
      description: Log a new schedule block.
      operationId: createCalendarEvent
      tags:
        - Calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /calendar/events/{id}:
    get:
      summary: Get calendar event details
      description: Retrieve details for a single schedule event.
      operationId: getCalendarEvent
      tags:
        - Calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update calendar event
      description: Modifies start, end, description, or trigger bounds.
      operationId: updateCalendarEvent
      tags:
        - Calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCalendarEventRequest'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete calendar event
      description: Removes calendar event schedule block.
      operationId: deleteCalendarEvent
      tags:
        - Calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived session access token passed in Authorization headers.
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Secure HTTP-only refresh token used to request new access tokens.
    TelegramSecretVerification:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Shared bot secret payload validating inbound Webhook calls.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Reserved API key token for external script integrations.
    OAuth2Placeholder:
      type: oauth2
      description: Reserved for future oauth integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.devmate.internal/oauth/authorize
          tokenUrl: https://auth.devmate.internal/oauth/token
          scopes:
            read: Grant view permissions on owned resources.
            write: Grant modification permissions on owned resources.
...
  headers:
    CorrelationIdResponseHeader:
      description: Ingress trace correlation tracker ID UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Input validation constraints violations.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationIdResponseHeader'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
...
    Unauthorized:
      description: Authentication failure due to missing or expired tokens.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimitError:
      description: Too many requests processed in window.
      headers:
        Retry-After:
          description: Wait seconds before retrying.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServerError:
      description: Internal system error log tracker.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    TelegramAuthRequest:
      type: object
      required:
        - id
        - first_name
        - auth_date
        - hash
      properties:
        id:
          type: integer
          description: Telegram user ID
        first_name:
          type: string
          description: Telegram user first name
        last_name:
          type: string
          description: Telegram user last name
        username:
          type: string
          description: Telegram username handle
        auth_date:
          type: integer
          description: Epoch timestamp of authorization
        hash:
          type: string
          pattern: '^[a-fA-F0-9]{64}$'
          description: Hex signature verifying verification data.
    AuthSuccessResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required:
            - accessToken
            - expiresIn
          properties:
            accessToken:
              type: string
            expiresIn:
              type: integer
              example: 3600
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    UpdateReminderRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        triggerTime:
          type: string
          format: date-time
        recurrence:
          type: string
    ReminderResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/ReminderDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ReminderDetail:
      type: object
      required:
        - id
        - text
        - triggerTime
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        triggerTime:
          type: string
          format: date-time
        recurrence:
          type: string
        status:
          $ref: '#/components/schemas/ReminderStatus'
        version:
          type: integer
...
    CreateCalendarEventRequest:
      type: object
      required:
        - title
        - startTime
        - endTime
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    UpdateCalendarEventRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    CalendarEventResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/CalendarEventDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    CalendarEventDetail:
      type: object
      required:
        - id
        - title
        - startTime
        - endTime
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    CalendarEventListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEventDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ResponseMeta:
      type: object
      required:
        - correlationId
        - traceId
      properties:
        correlationId:
          type: string
          format: uuid
        traceId:
          type: string
          format: uuid
...
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Request resource not found.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ValidationErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
            - details
          properties:
            code:
              type: string
              const: VALIDATION_FAILED
            message:
              type: string
              example: Invalid input parameters.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: object
                required:
                  - field
                  - issue
                properties:
                  field:
                    type: string
                    example: amount
                  issue:
                    type: string
                    example: Value must be a positive string.
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ReminderStatus:
      type: string
      enum: [PENDING, FIRED, MISSED, SNOOZED, PAUSED, COMPLETED]
...